rgw/rgw_op: cast size of copied chunk to uint64_t
after yaml-to-cxx migration, rgw_copy_obj_progress_every_bytes will be
an uint64_t. so cast the left-hand operator to `uint64_t` before
comparing them.

+++ src/rgw/rgw_op.cc
@@ -5053,7 +5053,8 @@ void RGWCopyObj::progress_cb(off_t ofs)
   if (!s->cct->_conf->rgw_copy_obj_progress)
     return;
 
-  if (ofs - last_ofs < s->cct->_conf->rgw_copy_obj_progress_every_bytes)
+  if (static_cast<uint64_t>(ofs - last_ofs) <
+      s->cct->_conf->rgw_copy_obj_progress_every_bytes)
     return;
 
   send_partial_response(ofs);
